// src/app/post.interface.ts
export interface PostInterface {
    id: string;
    channelname: string;
    title: string;
    high_thumbnail: string;
    low_thumbnail: string;
    medium_thumbnail: string;
    tags: string[];
};

export interface PostResponseInterface {
    id: string;
    channelname: string;
    title: string;
    "high_thumbnail": string;
    "low_thumbnail": string;
    "medium thumbnail": string;
    tags: string[];
};